requirements.txt
aiofiles>=0.6.0
aiohttp>=3.6.2
discord.py>=1.6.0
pymysql>=1.0.2
